# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduConnect EdTech Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (super admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /support:
    get:
      operationId: listSupportTickets
      tags: [support]
      summary: List support tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved]
      responses:
        "200":
          description: Support tickets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createSupportTicket
      tags: [support]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportTicketBody"
      responses:
        "200":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /support/{id}/respond:
    patch:
      operationId: respondToTicket
      tags: [support]
      summary: Admin responds to a support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondToTicketBody"
      responses:
        "200":
          description: Response saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin, super_admin, planner]
        status:
          type: string
          enum: [pending, approved, rejected, active]
        phone:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - fullName
        - email
        - role
        - status
        - createdAt
...
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved]
        adminResponse:
          type: ["string", "null"]
        respondedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - studentName
        - subject
        - message
        - status
        - createdAt
        - updatedAt
    CreateSupportTicketBody:
      type: object
      properties:
        subject:
          type: string
        message:
          type: string
      required:
        - subject
        - message
    RespondToTicketBody:
      type: object
      properties:
        adminResponse:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved]
      required:
        - adminResponse
        - status
...